feat: clone to opt and use vim.cmd to clone instead of vim.system
since the cloning must happen synchronously due to how baggage is
intended to be used, it was necessary to call system with sync = true.
however that would block the UI, which resultet in initial launches
being blank for the duration of all the clones.

vim.cmd ALSO runs synchronously however it will print the stdout to the
screen, so that the user would get at least some indication of what is
going on.

moving packages to opt, so it's possible to load only those that are
actually needed (and start can be used for baggage itself in the next
step).